chore(front): extract voice recording and transcription in a dedicated service
